+++ src/ServiceCollection.ts
@@ -1,11 +1,11 @@
 import { ServiceProvider } from './ServiceProvider';
-import type { SourceType, InstanceFactory, ServiceBuilder, ServiceDescriptor, ServiceIdentifier, ServiceImplementation, Newable } from './types';
+import type { SourceType, InstanceFactory, ServiceBuilder, ServiceDescriptor, ServiceIdentifier, ServiceImplementation, Newable, ServiceCollectionOptions } from './types';
 import type { IServiceCollection, IServiceModule, IServiceProvider } from './interfaces';
 import { Lifetime } from './constants';
 import { log } from './debug';
 
 export class ServiceCollection implements IServiceCollection {
-  constructor(private readonly services = new Map<ServiceIdentifier<any>, ServiceDescriptor<any>[]>()) {
+  constructor(public readonly options: ServiceCollectionOptions, private readonly services = new Map<ServiceIdentifier<any>, ServiceDescriptor<any>[]>()) {
   }
   
   public registerModules(modules: Newable<IServiceModule>[]): void {
@@ -52,7 +52,7 @@ export class ServiceCollection implements IServiceCollection {
       clonedMap.set(key, clonedDescriptors);
     }
 
-    return new ServiceCollection(clonedMap);
+    return new ServiceCollection(this.options, clonedMap);
   }
 
   public buildProvider(): IServiceProvider {